Fix issue #11086275: Thumbnail only created once for top activity
If the last screenshot activity is resumed, we need to always capture
a new screenshot, because it can change at any time.

On the other hand, never create a thumbnail for tasks that have set
themselves to not show on the recent tasks lists, since we have no
use for them.

Change-Id: I38523afc966c125da93339e0100da950119cdf99