NotificationPanel now shows correct date
There was a problem with a cached instance of a Date object such that it
would retain information about the default TimeZone that was not updated
when the TimeZone was set to the correct value. This resulted in NotificationPanel
sometimes showing the wrong date, based on the time and the delta between
the local timezone and GMT.

The fix is to use an ACTION_TIMEZONE_CHANGED broadcast to recreate the cached
Date object appropriately.

Issue #10854196 Incorrect date showing in notification shade

Change-Id: Ibf883e1ff37f4d0aae6840fb19e43e7e9e45bcc2