Use dialog animations for dialog style.
Using activity animations on non-fullscreen activities that sit over
a wallpaper activity cause the exiting activity to show up underneath
the next activity. If the next activity happens to be a wallpaper
activity then this will be janky as the exiting activity will go from
on top of the incoming activity to on top of the wallpaper as it
animates away.

Fixes bug 11200253.

Change-Id: I7e81c50cfe7ae106da2d62f5866f43fc046181c5