Temporary detached views not properly reported to accessibility layer.
A view may be temporary detached. While in this state the view
state may change for which we fire accessibility events. Clients
use them to know when the view changed and what the change is.
However, detached views do not fire accessibility events.
Hence, a client may hold a stale AccessibilityNodeInfo since
the changes to its source view happened in a temporary detached
state.

bug:11388752

Change-Id: I9052700ca9195841cc2881b65c17f5615e6f71fb