Support dynamic rotation of HDMI displays
To support "giant" phones which are really just normal phones
strapped to external HDMI displays, add the property
persist.demo.hdmirotates.  It defaults to false, but when it is
set to true, the FLAG_ROTATES_WITH_CONTENT to set on the
display.  This allows the external display to show the same
display as the built-in display as the "giant" phone display is
rotated.

Note that previously, FLAG_ROTATES_WITH_CONTENT was only set on
the built-in display.  The code that checked the flag also
explicitly ignored it on any display that was not the built-in
display.  This added check was removed to allow the flag to be
functional on other displays.

Change-Id: I55b249140b1f61fb98cac586f7e4d48e2f5b3e30
Signed-off-by: Scott Anderson <saa@android.com>