Work around nested SaveLayerAlpha bug
bug:11481077

Because TextView marquee and KeyguardSecurityViewFlipper use SaveLayer
(for the marquee fading effect and view alpha, respectively), both can
be nested, which hits a bug in the renderer resulting in artifacts.

Workaround this by setting a hardware layer on the marquee, so that
those marquee effects are drawn outside of the ViewFlipper.

Change-Id: If8b7ebb20d3959cb930a8bea2146ba3f8e659612